Pertemuan 13

+++ system/app/Models/Traits/Attributes/ProdukAttributes.php
@@ -18,7 +18,7 @@ trait ProdukAttributes {
 			$foto = request()->file('foto');
 			$destination = "images/produk";
 			$randomStr = Str::random(5);
-			$filename = $this->id."-".time()."-".$randomStr.".".$foto->extension();
+			$filename = $this->uuid."-".time()."-".$randomStr.".".$foto->extension();
 			$url = $foto->storeAs($destination, $filename);
 			$this->foto = "app/".$url;
 			$this->save();
@@ -27,11 +27,12 @@ trait ProdukAttributes {
 
 	function handleDelete(){
 		$foto = $this->foto;
-		$path = public_path($foto);
-		if(file_exists($path)){
-			unlink($path);
+		if($foto){
+			$path = public_path($foto);
+			if(file_exists($path)){
+				unlink($path);
+			}
+			return true;
 		}
-		return true;
-		
 	}
 }
